feat(#126): Phase 275 — DRY consolidation (zone parser + helper extraction)
Extract parseZoneContent() to lib/muninn-helpers.ts, eliminating duplicate
in zone-history route and use-observations hook. Move pure observation
metric helpers to lib/observation-helpers.ts.
